some more improvements following reviews [skip ci]
