+++ frontend_lib/package.json
@@ -42,6 +42,7 @@
     "react-device-detect": "^1.7.5",
     "react-dropzone": "^7.0.1",
     "react-i18next": "7.4.0",
+    "react-onclickoutside": "^6.9.0",
     "react-router-dom": "^5.2.0",
     "regenerator-runtime": "^0.13.2",
     "uuid": "^8.1.0"
